Moved EditableTextContent to core-components.
