+++ app/core-components/EditableTextContent/index.test.js
@@ -3,7 +3,7 @@
 import * as React from 'react';
 import { mount, shallow } from 'enzyme';
 
-import EditableTextContent from './EditableTextContent';
+import EditableTextContent from '.';
 
 describe(`EditableTextContent`, (): void => {
