+++ app/modules/contentItems/components/EditableDisplay/types/Paragraph.js
@@ -4,12 +4,12 @@ import _ from 'lodash';
 import * as React from 'react';
 
 import type { Identifier } from 'types/model';
+import EditableTextContent from 'core-components/EditableTextContent';
 
 import type { ParagraphContentItem } from '../../../model';
 
 import { passThroughProps } from '..';
 import DisplayBlockWrapper from '../DisplayBlockWrapper';
-import EditableTextContent from '../EditableTextContent';
 
 type PassedProps = {
   contentItem: ParagraphContentItem,
